Match arbitrary dates when "today" key is used in a search format pattern
