+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `--patch` as command line options to display the planned changes as a patch instead of performing
     any operations.
 * Multi-line search and replacement patterns.
+* Match arbitrary dates when `today` key is used in a search format pattern.
 
 ### Changed
