+++ hyper_bump_it/_hyper_bump_it/config/core.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from ..text_formatter import keys
+from ..format_pattern import keys
 
 
 class GitAction(str, Enum):
