--- hyper_bump_it/_hyper_bump_it/text_formatter/__init__.py
@@ -1,4 +1,0 @@
-from . import keys as keys
-from .text_formatter import FormatContext, TextFormatter
-
-__all__ = ["keys", "TextFormatter", "FormatContext"]
