+++ requirements-test.txt
@@ -6,5 +6,6 @@ isort==5.12.0
 mypy==1.0.1
 pytest==7.2.1
 pytest-cov==4.0.0
+pytest-freezer==0.4.6
 pytest-mock==3.10.0
 pdbpp
